Modellazione della virtual machine

+++ src/Chip8.Model/Sprites/ISprite.cs
@@ -2,7 +2,7 @@ namespace Chip8.Model.Sprites;
 
 public interface ISprite
 {
-    int Width { get; }
-    int Height { get; }
-    bool this[int x, int y] { get; }
+    byte Width { get; }
+    byte Height { get; }
+    bool this[byte x, byte y] { get; }
 }
